+++ src/RcppR6.cpp
@@ -890,15 +890,6 @@ void Leaf__z_soil_mid___set(phylloptim::RcppR6::RcppR6<phylloptim::Leaf> obj_, s
 }
 
 // [[Rcpp::export]]
-double Leaf__dz___get(phylloptim::RcppR6::RcppR6<phylloptim::Leaf> obj_) {
-  return obj_->roots_.dz_;
-}
-// [[Rcpp::export]]
-void Leaf__dz___set(phylloptim::RcppR6::RcppR6<phylloptim::Leaf> obj_, double value) {
-  obj_->roots_.dz_ = value;
-}
-
-// [[Rcpp::export]]
 int Leaf__soil_number_of_depths___get(phylloptim::RcppR6::RcppR6<phylloptim::Leaf> obj_) {
   return obj_->roots_.soil_number_of_depths_;
 }